refactor: split popup logic into separate components
