+++ src/components/crm/AddCardForm/index.tsx
@@ -23,7 +23,7 @@ import {
 } from "../AddCardCrm/helpers/locations";
 import { createLocation } from "@/src/utils/locations";
 import { sortDiagnosesOrProcedures } from "../AddCardCrm/helpers/sort";
-import { Location, defaultValues } from "../AddCardCrm/types/types";
+import { AnimalCard, Location, defaultValues } from "../AddCardCrm/types/types";
 
 const API_CRM_PATH = process.env.NEXT_PUBLIC_API_CRM_PATH;
 const API_ANIMALS_PATH = process.env.NEXT_PUBLIC_API_ANIMALS_PATH;
@@ -44,7 +44,6 @@ const AddCardForm = () => {
     control,
     handleSubmit,
     trigger,
-    getValues,
     formState: { isValid, isSubmitted },
   } = methods;
 
@@ -87,16 +86,14 @@ const AddCardForm = () => {
         death__comment: null,
       };
 
-      console.log(updatedData);
+      const result = await post<AnimalCard>(
+        `${API_CRM_PATH}${API_ANIMALS_PATH}`,
+        updatedData
+      );
 
-      // const result = await post<AnimalCard>(
-      //   `${API_CRM_PATH}${API_ANIMALS_PATH}`,
-      //   updatedData
-      // );
-
-      // if (result?.id) {
-      //   window.location.href = "/crm/catalog";
-      // }
+      if (result?.id) {
+        window.location.href = "/crm/catalog";
+      }
     } catch (error) {
       throw error;
     } finally {
